refactor: Add Codolio social icon to Contact page
The code changes in Contact.jsx and Contact.css add a Codolio social icon to the Contact page. This refactor includes importing the Codolio SVG file and adding the icon to the social links section. This change enhances the social media presence and provides an additional way for users to connect with the developer.